requirements.txt
Adafruit-GPIO==1.0.3
Adafruit-PureIO==1.1.5
astroid==2.4.2
autopep8==1.5.3
certifi==2020.6.20
chardet==3.0.4
ecdsa==0.15
esptool==2.8
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pyaes==1.6.1
pycodestyle==2.6.0
pylint==2.5.3
pyserial==3.4
python-statemachine==0.8.0
requests==2.24.0
RPi.GPIO==0.7.0
six==1.15.0
spidev==3.5
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.12.1
